requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
docker==7.1.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.2.3
PyYAML~=6.0.2